Adding new 8 channel version of Power Meter
As the mux is 8 channels, why limit ourselves to just 3 CTs? :) To keep
the small footprint of the original board design as close as possible,
the board was increased from 2 layers to 4 layers. The signal lines
going from the mux to the CT pin header uses the inner two layers of the
board design.

This new design passes KiCad validation checks, but has not been
reviewed by the original author or anyone else besides me and also
hasn't had a pre-production test run completed yet either.

Definitely USE AT YOUR OWN RISK! ;)